2018.12.25

+++ chapter15/ex_15_03.cpp
@@ -4,7 +4,7 @@
 
 #include "ex_15_03.h"
 
-double PrintTotal(std::ostream &os, const Quote &item, std::int32_t n) {
+double PrintTotal(std::ostream &os, const Quote &item, std::size_t n) {
   double ret{item.NetPrice(n)};
   os << "ISBN: " << item.Isbn()
      << " # sold: " << n << " total due: " << ret << '\n';
